// amplify/backend/types/amplify-dependent-resources-ref.d.ts
export type AmplifyDependentResourcesAttributes = {
    "hosting": {
        "S3AndCloudFront": {
            "Region": "string",
            "HostingBucketName": "string",
            "WebsiteURL": "string",
            "S3BucketSecureURL": "string",
            "CloudFrontDistributionID": "string",
            "CloudFrontDomainName": "string",
            "CloudFrontSecureURL": "string",
            "CloudFrontOriginAccessIdentity": "string"
        }
    },
    "custom": {
        "resource": {
            "PublicKey": "string",
            "trustedKeyGroupId": "string"
        }
    }
}